fix: Prevent wrong address attestation from being used
When UN verification is used with ZKProof, and the identifier
attestation is signed with a different address compared to
the UN, verification will fail. This fix sends the ethereum
address of the connected account to attestation.id so
attestation.id doesn't return a previously generated attestation
with the wrong address.