fix: Add deep-merge of config with defaults
This fix adds deep-merge of config into defaultConfig to allow partial
specification of nested options and the ability to disable option like
userAgentSupported.full.config.

This adds a null propagation to usages of isUserAgentSupported and a nullish
check in the function to allow disabling by setting these config values
to null or undefined.