fix: Disable third-party cookie check to fix dapp-browsers TKN-252
The third party cookie check requires that the ticket is created in
the outlet. This creates a special localStorage flag that can be checked
via negotiator to infer cross-origin local storage access.

The problem with this is that if the check fails, a popup is used automatically.
This may happen when the ticket is not issued or on the ticket issuer page itself.
Alphawallet android and many other dapp browsers do not support popups and instead
navigate to the page in the same window. This causes a white screen with no way to return.

This commit disables the check and only forces the use of tabs in browsers known to restrict
iframe access to local storage, like safafi:
https://stackoverflow.com/questions/63922558/safari-localstorage-not-shared-between-iframes-hosted-on-same-domain